break up metron.go
